Extract Summary preview galleries #290
